Implemented test mode
Removed verbose option from help message because the
output would be so voluminous that no one could follow it.
It is better to run the program in test mode, read that output after
e.g. having saved it to a file and check the proposed actions for
errors (so you have a chance to fix your tags)

Until now the output comes from the TagWalker class.
This is likely to change in the next few commits