Completely remove "verbose" mode and add "cleanup" mode
Cleanup mode deletes empty directories inside the directory tree.
If the flag is checked, sync() is called frequently to determine whether
a directory is empty. This is needed due to slow IO. Should be mentioned
in the manpage.
Also, the output in main() now gives info about deleted dirs (if flag
was checked) and unhandleable files